Remove obsolete LONG_USAGE which breaks xgettext
The obsolete LONG_USAGE variable has the following message in it:

    A'\''--B'\''--C'\''

And such complex LONG_USAGE message will breaks xgettext when extract
l10n messages. But if remove single quotes from the message, xgettext
works fine on 'git-rebase.sh'.

Since there is a mordern OPTIONS_SPEC variable in use in this script,
it's safe to remove the obsolte USAGE and LONG_USAGE variables.

Signed-off-by: Jiang Xin <worldhello.net@gmail.com>